squash! feat: implement history interface.
Notes: Wrap history listener in a class so that instances can be enabled
and disabled from the class interface. Then, change events posted from
the undo/redo actin can be ignored by disabling the listener
momentarily. Also store the listener so that it can be deregistered
later, since anonymous delegates cannot be unsubscribed.